@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 from typing import TypeVar
-from typing import Union
 
 from pyvista.core.composite import MultiBlock
 from pyvista.core.dataobject import DataObject
@@ -26,7 +25,7 @@ _PointSetType.__doc__ = """Type variable for PyVista ``PointSet`` classes."""
 _DataSetType = TypeVar('_DataSetType', bound=DataSet)
 _DataSetType.__doc__ = """Type variable for :class:`~pyvista.DataSet` classes."""
 
-_DataSetOrMultiBlockType = TypeVar('_DataSetOrMultiBlockType', bound=Union[DataSet, MultiBlock])
+_DataSetOrMultiBlockType = TypeVar('_DataSetOrMultiBlockType', bound=DataSet | MultiBlock)
 _DataSetOrMultiBlockType.__doc__ = (
     """Type variable for :class:`~pyvista.DataSet` or :class:`~pyvista.MultiBlock` classes."""
 )
