@@ -18,7 +18,7 @@ from pyvista.core.utilities.helpers import wrap
 from pyvista.core.utilities.misc import abstract_class
 
 if TYPE_CHECKING:
-    from typing import Callable
+    from collections.abc import Callable
 
     from pyvista import MultiBlock
     from pyvista.core.composite import _TypeMultiBlockLeaf
