@@ -162,8 +162,11 @@ f, ax = plt.subplots(
 alphas = [0.5 + i for i in range(5)]
 betas = [*reversed(alphas)]
 N = int(1e4)
-data = [rng.beta(alpha, beta, N) for alpha, beta in zip(alphas, betas)]
-labels = [f'$\\alpha={alpha:.1f}\\,;\\,\\beta={beta:.1f}$' for alpha, beta in zip(alphas, betas)]
+data = [rng.beta(alpha, beta, N) for alpha, beta in zip(alphas, betas, strict=False)]
+labels = [
+    f'$\\alpha={alpha:.1f}\\,;\\,\\beta={beta:.1f}$'
+    for alpha, beta in zip(alphas, betas, strict=False)
+]
 ax.violinplot(data)
 ax.set_xticks(np.arange(1, 1 + len(labels)))
 ax.set_xticklabels(labels)
