@@ -36,11 +36,11 @@ if TYPE_CHECKING:
 # get the int type from vtk
 ID_TYPE: type[np.int32 | np.int64] = _get_vtk_id_type()
 
-# determine if using at least vtk 9.0.0
-if vtk_version_info.major < 9:  # pragma: no cover
+# determine if using at least vtk 9.1.0
+if vtk_version_info < (9, 1, 0):  # pragma: no cover
     from pyvista.core.errors import VTKVersionError
 
-    msg = 'VTK version must be 9.0.0 or greater.'
+    msg = 'VTK version must be 9.1.0 or greater.'
     raise VTKVersionError(msg)
 
 # catch annoying numpy/vtk future warning:
