@@ -6,13 +6,15 @@ from inspect import Parameter
 from inspect import Signature
 import os
 from pathlib import Path
-from typing import Callable
+from typing import TYPE_CHECKING
 from typing import TypeVar
 from typing import overload
 import warnings
 
 from typing_extensions import ParamSpec
 
+if TYPE_CHECKING:
+    from collections.abc import Callable
 from pyvista._version import version_info
 
 _MAX_POSITIONAL_ARGS = 3  # Should match value in pyproject.toml
