@@ -3224,7 +3224,7 @@ class AxesGeometrySource(_NoNewAttrMixin):
         # Init datasets
         names = ['x_shaft', 'y_shaft', 'z_shaft', 'x_tip', 'y_tip', 'z_tip']
         polys = [pyvista.PolyData() for _ in range(len(names))]
-        self._output = pyvista.MultiBlock(dict(zip(names, polys)))
+        self._output = pyvista.MultiBlock(dict(zip(names, polys, strict=False)))
 
         # Store shaft/tip references in separate vars for convenience
         self._shaft_datasets = (polys[0], polys[1], polys[2])
@@ -3824,7 +3824,7 @@ class OrthogonalPlanesSource(_NoNewAttrMixin):
         self._normal_sign = tuple(valid_sign)
 
         # Modify sources
-        for source, axis_vector, sign_ in zip(self.sources, np.eye(3), valid_sign):
+        for source, axis_vector, sign_ in zip(self.sources, np.eye(3), valid_sign, strict=False):
             has_positive_normal = np.dot(source.normal, axis_vector) > 0
             if has_positive_normal and sign_ == '-':
                 source.flip_normal()
@@ -3927,12 +3927,12 @@ class OrthogonalPlanesSource(_NoNewAttrMixin):
             dtype_out=float,
             to_tuple=True,
         )
-        for source, dist in zip(self.sources, valid_distance):
+        for source, dist in zip(self.sources, valid_distance, strict=False):
             source.push(dist)
 
     def update(self: OrthogonalPlanesSource) -> None:
         """Update the output of the source."""
-        for source, plane in zip(self.sources, self._output):
+        for source, plane in zip(self.sources, self._output, strict=False):
             plane.copy_from(source.output)
 
     @property
@@ -4391,7 +4391,7 @@ class CubeFacesSource(CubeSource):
         output = self._output
 
         # Modify each face mesh of the output
-        for index, (name, points) in enumerate(zip(self.names, face_points)):
+        for index, (name, points) in enumerate(zip(self.names, face_points, strict=False)):
             output.set_block_name(index, name)
             face_poly = output[index]
             face_center = np.mean(points, axis=0)
