@@ -64,7 +64,9 @@ def pytest_generate_tests(metafunc):
                 return any(base.__name__[:3].lower() == 'vtk' for base in bases)
 
             inherits_from_vtk = {
-                name: cls for name, cls in zip(class_names, class_types) if inherits_from_vtk(cls)
+                name: cls
+                for name, cls in zip(class_names, class_types, strict=False)
+                if inherits_from_vtk(cls)
             }
             assert inherits_from_vtk
             return inherits_from_vtk
@@ -84,7 +86,7 @@ def pytest_generate_tests(metafunc):
 
         class_map = {
             name: cls
-            for name, cls in zip(class_names, class_types)
+            for name, cls in zip(class_names, class_types, strict=False)
             if not name.startswith('_') and not issubclass(cls, tuple(SKIP_SUBCLASS))
         }
         metafunc.parametrize('pyvista_class', list(class_map.values()), ids=list(class_map.keys()))
